requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
pydantic>=2.6.0
pydantic-settings>=2.1.0
slowapi>=0.1.9
httpx>=0.26.0
python-docx>=1.1.0
python-dotenv>=1.0.1
pytesseract>=0.3.10
Pillow>=10.0.0
